feat : market detail userInfo 수정
